Add methods to create dicts inside models. Fix client methods. Fix time check fro drivers and clients. Fix travel add time. Fix travel search filter. Fix register-form bad indentation. Updated travel-list in driver

+++ map/models.py
@@ -2,8 +2,8 @@ from django.db import models
 
 # Create your models here.
 class PosLatLng(models.Model):
-    lat=models.PositiveSmallIntegerField()
-    lng=models.PositiveSmallIntegerField()
+    lat=models.FloatField()
+    lng=models.FloatField()
 
     def __str__(self):
         return self.lat.__str__() + ", " + self.lng.__str__()
